dayman progress 3

+++ shared/src/commonMain/kotlin/com/romankozak/forwardappmobile/shared/features/daymanagement/data/DayTaskMapper.kt
@@ -17,7 +17,7 @@ fun DayTasks.toDomain(): DayTask {
         recurringTaskId = this.recurringTaskId,
         taskType = this.taskType,
         entityId = this.entityId,
-        order = this.order.toInt(),
+        order = this.order,
         priority = this.priority,
         status = this.status,
         completed = this.completed != 0L,
